Move Avx2 GetColSums and GetAbsDxColSums to SimdAvx2ColSums.cpp
Extract Avx2::GetColSums and Avx2::GetAbsDxColSums from
SimdAvx2Statistic.cpp into a dedicated SimdAvx2ColSums.cpp, and
register the new file in the VS2022 Avx2 project.
