+++ prj/vs2022/Avx2.vcxproj.filters
@@ -190,6 +190,9 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx2StatisticMoments.cpp">
       <Filter>Avx2\Statistics</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Simd\SimdAvx2ColSums.cpp">
+      <Filter>Avx2\Statistics</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Simd\SimdAvx2SquaredDifferenceSum.cpp">
       <Filter>Avx2\Statistics</Filter>
     </ClCompile>
